Sol Ver1.4.2
EventController 삭제
이벤트는 메인뷰에서 선언
