+++ src/com/aroundThirty/Resource/BR.java
@@ -1,6 +1,5 @@
 package com.aroundThirty.Resource;
 
-import com.aroundThirty.Controller.EventController;
 import com.aroundThirty.model.ReportDto;
 
 public class BR {
