+++ src/com/aroundThirty/View/AdoptPage.java
@@ -1,6 +1,6 @@
 package com.aroundThirty.View;
 
-import com.aroundThirty.Controller.EventController;
+import com.aroundThirty.Controller.MainController;
 
 import javax.swing.*;
 import java.awt.*;
@@ -55,7 +55,7 @@ public class AdoptPage extends JPanel {
 
                 }
             });
-            EventController.onClick();
+            MainController.onClick();
         }
         JPanel buttonPane = new JPanel();
         JButton btn1 = new JButton("1");
@@ -67,7 +67,7 @@ public class AdoptPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("입양동물 Page 01 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -85,7 +85,7 @@ public class AdoptPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("입양동물 Page 02 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -102,7 +102,7 @@ public class AdoptPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("입양동물 Page 03 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -119,7 +119,7 @@ public class AdoptPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("입양동물 Page 04 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -136,7 +136,7 @@ public class AdoptPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("입양동물 Page 05 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -153,7 +153,7 @@ public class AdoptPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("입양동물 Page 06 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -170,7 +170,7 @@ public class AdoptPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("입양동물 Page 07 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -187,7 +187,7 @@ public class AdoptPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("입양동물 Page 08 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -204,7 +204,7 @@ public class AdoptPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("입양동물 Page 09 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
