+++ src/com/aroundThirty/View/MissingPage.java
@@ -1,6 +1,6 @@
 package com.aroundThirty.View;
 
-import com.aroundThirty.Controller.EventController;
+import com.aroundThirty.Controller.MainController;
 
 import javax.swing.*;
 import java.awt.*;
@@ -56,7 +56,7 @@ public class MissingPage extends JPanel {
 
                 }
             });
-            EventController.onClick();
+            MainController.onClick();
         }
         JPanel buttonPane = new JPanel();
         JButton btn1 = new JButton("1");
@@ -68,7 +68,7 @@ public class MissingPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("실종동물 Page 01 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -86,7 +86,7 @@ public class MissingPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("실종동물 Page 02 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -103,7 +103,7 @@ public class MissingPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("실종동물 Page 03 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -120,7 +120,7 @@ public class MissingPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("실종동물 Page 04 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -137,7 +137,7 @@ public class MissingPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("실종동물 Page 05 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -154,7 +154,7 @@ public class MissingPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("실종동물 Page 06 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -171,7 +171,7 @@ public class MissingPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("실종동물 Page 07 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -188,7 +188,7 @@ public class MissingPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("실종동물 Page 08 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
@@ -205,7 +205,7 @@ public class MissingPage extends JPanel {
                     mainPanel.add(rbtn = new JButton("실종동물 Page 09 " + i));
                 }
                 rbtn.setPreferredSize(new Dimension(200, 300));
-                EventController.onClick();
+                MainController.onClick();
 
                 revalidate();
                 repaint();
